+++ CHANGELOG.md
@@ -14,6 +14,7 @@
 
 
 ### 🔧 Other Changes
-* 2025-03-29 - check os name
+* 2025-04-25 - usage of loguru.logger
+* 2025-03-29 - check os name [skip ci]
 * 2025-03-20 - selenium test [skip ci]
 * 2025-03-09 - asyncio lock [skip ci]